feat(auth): Implement authentication module and refactor user services
Added AuthModule with its controller (AuthController) and service (AuthService) to handle admin and user registration.

Moved CreateAdminDto and CreateUserSelfRegistrationDto to the authentication module, renaming them as RegisterAdminDto and RegisterUserDto.

Refactored UsersService to remove registration and hashing logic, delegating it to AuthService.

Updated UsersModule to export UsersService for use in other modules.

Improved consistency in ID handling by changing the type from number to string in ProjectsController and ProjectsService methods.

Simplified CreateUserDto and added support for roles and admin relationships.